moved test_zipkin_doesnt_spew_on_first_log to tests/logging_helper_test.py
