+++ tests/logging_helper_test.py
@@ -7,6 +7,23 @@ from pyramid_zipkin.exception import ZipkinError
 from pyramid_zipkin.request_helper import ZipkinAttrs
 
 
+# This test _must_ be the first test in this file
+def test_zipkin_doesnt_spew_on_first_log(capfd):
+    import logging
+
+    zipkin_logger = logging.getLogger('pyramid_zipkin.logger')
+
+    zipkin_logger.debug({
+        'annotations': {'foo': 2},
+        'name': 'bar',
+    })
+
+    out, err = capfd.readouterr()
+
+    assert not err
+    assert not out
+
+
 @pytest.fixture
 def context():
     attr = ZipkinAttrs(None, None, None, None, False)
@@ -239,7 +256,9 @@ def test_log_span_uses_default_stream_name_if_not_provided(thrift_obj, create_sp
         '0000000000000002', '0000000000000001', '00000000000000015',
         'span', 'ann', 'binary_ann', registry
     )
-    transport_handler.assert_called_once_with('zipkin', thrift_obj.return_value)
+    transport_handler.assert_called_once_with(
+        'zipkin', thrift_obj.return_value
+    )
 
 
 @mock.patch('pyramid_zipkin.logging_helper.create_span', autospec=True)
